Move OpenTelemetry instrumentation to right package
