+++ dd-java-agent/instrumentation/opentelemetry/opentelemetry-1.4/src/main/java/datadog/trace/instrumentation/opentelemetry14/OpenTelemetryInstrumentation.java
@@ -1,4 +1,4 @@
-package datadog.trace.opentelemetry14;
+package datadog.trace.instrumentation.opentelemetry14;
 
 import static datadog.trace.agent.tooling.bytebuddy.matcher.HierarchyMatchers.implementsInterface;
 import static datadog.trace.agent.tooling.bytebuddy.matcher.NameMatchers.named;
@@ -8,20 +8,15 @@ import static net.bytebuddy.matcher.ElementMatchers.takesNoArguments;
 
 import com.google.auto.service.AutoService;
 import datadog.trace.agent.tooling.Instrumenter;
-import datadog.trace.bootstrap.instrumentation.api.AgentSpan;
-import datadog.trace.bootstrap.instrumentation.api.AgentTracer;
-import datadog.trace.bootstrap.instrumentation.api.AttachableWrapper;
-import io.opentelemetry.api.trace.Span;
-import io.opentelemetry.context.Context;
+import io.opentelemetry.api.trace.TracerProvider;
 import net.bytebuddy.asm.Advice;
 import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.matcher.ElementMatcher;
 
 @AutoService(Instrumenter.class)
-public class OpenTelemetryContextInstrumentation extends Instrumenter.Tracing
+public class OpenTelemetryInstrumentation extends Instrumenter.Tracing
     implements Instrumenter.CanShortcutTypeMatching {
-
-  public OpenTelemetryContextInstrumentation() {
+  public OpenTelemetryInstrumentation() {
     super("opentelemetry.experimental", "opentelemetry-1");
   }
 
@@ -32,7 +27,7 @@ public class OpenTelemetryContextInstrumentation extends Instrumenter.Tracing
 
   @Override
   public String hierarchyMarkerType() {
-    return "io.opentelemetry.context.ContextStorage";
+    return "io.opentelemetry.api.OpenTelemetry";
   }
 
   @Override
@@ -43,8 +38,8 @@ public class OpenTelemetryContextInstrumentation extends Instrumenter.Tracing
   @Override
   public String[] knownMatchingTypes() {
     return new String[] {
-      "io.opentelemetry.context.ThreadLocalContextStorage",
-      "io.opentelemetry.context.StrictContextStorage",
+      "io.opentelemetry.api.DefaultOpenTelemetry",
+      "io.opentelemetry.api.GlobalOpenTelemetry$ObfuscatedOpenTelemetry"
     };
   }
 
@@ -62,6 +57,7 @@ public class OpenTelemetryContextInstrumentation extends Instrumenter.Tracing
       packageName + ".OtelSpan$NoopSpan",
       packageName + ".OtelSpan$NoopSpanContext",
       packageName + ".OtelSpanBuilder",
+      packageName + ".OtelSpanBuilder$1",
       packageName + ".OtelSpanContext",
       packageName + ".OtelTracer",
       packageName + ".OtelTracerBuilder",
@@ -71,30 +67,19 @@ public class OpenTelemetryContextInstrumentation extends Instrumenter.Tracing
 
   @Override
   public void adviceTransformations(AdviceTransformation transformation) {
-    // Context.current()
+    // TracerProvider.getTracerProvider()
     transformation.applyAdvice(
         isMethod()
-            .and(named("current"))
+            .and(named("getTracerProvider"))
             .and(takesNoArguments())
-            .and(returns(named("io.opentelemetry.context.Context"))),
-        OpenTelemetryContextInstrumentation.class.getName() + "$ContextCurrentAdvice");
+            .and(returns(named("io.opentelemetry.api.trace.TracerProvider"))),
+        OpenTelemetryInstrumentation.class.getName() + "$TracerProviderAdvice");
   }
 
-  public static class ContextCurrentAdvice {
+  public static class TracerProviderAdvice {
     @Advice.OnMethodExit(suppress = Throwable.class)
-    public static void current(@Advice.Return(readOnly = false) Context result) {
-      AgentSpan agentSpan = AgentTracer.activeSpan();
-      Span otelSpan = null;
-      if (agentSpan instanceof AttachableWrapper) {
-        Object wrapper = ((AttachableWrapper) agentSpan).getWrapper();
-        if (wrapper instanceof OtelSpan) {
-          otelSpan = (OtelSpan) wrapper;
-        }
-      }
-      if (otelSpan == null) {
-        otelSpan = agentSpan == null ? OtelSpan.invalid() : new OtelSpan(agentSpan);
-      }
-      result = new OtelContext(otelSpan);
+    public static void returnProvider(@Advice.Return(readOnly = false) TracerProvider result) {
+      result = OtelTracerProvider.INSTANCE;
     }
   }
 }
